// functions/src/types/line-events.ts
export interface LineMessageEvent {
  type: 'message';
  mode: 'active' | 'standby';
  timestamp: number;
  source: LineUser | LineGroup | LineRoom;
  webhookEventId: string;
  deliveryContext: {
    isRedelivery: boolean;
  };
  replyToken: string;
  message: LineMessage;
}
// ...
export type LineMessage =
  | LineTextMessage
  | LineStickerMessage
  | LineImageMessage
  | LineVideoMessage
  | LineAudioMessage
  | LineFileMessage
  | LineLocationMessage;

export interface LineTextMessage {
  id: string;
  type: 'text';
  quoteToken?: string;
  text: string;
  emojis?: LineEmoji[];
  mention?: LineMention;
}

export interface LineStickerMessage {
  id: string;
  type: 'sticker';
  quoteToken?: string;
  packageId: string;
  stickerId: string;
  stickerResourceType:
    | 'STATIC'
    | 'ANIMATION'
    | 'SOUND'
    | 'ANIMATION_SOUND'
    | 'POPUP'
    | 'POPUP_SOUND'
    | 'CUSTOM'
    | 'MESSAGE';
  keywords?: string[];
  text?: string;
}

export interface LineImageMessage {
  id: string;
  type: 'image';
  quoteToken?: string;
  contentProvider?: {
    type: 'line' | 'external';
    originalContentUrl?: string;
    previewImageUrl?: string;
  };
  imageSet?: {
    id: string;
    index: number;
    total: number;
  };
}

export interface LineVideoMessage {
  id: string;
  type: 'video';
  quoteToken?: string;
  duration: number;
  contentProvider?: {
    type: 'line' | 'external';
    originalContentUrl?: string;
    previewImageUrl?: string;
  };
}

export interface LineAudioMessage {
  id: string;
  type: 'audio';
  quoteToken?: string;
  duration: number;
  contentProvider?: {
    type: 'line' | 'external';
    originalContentUrl?: string;
  };
}

export interface LineFileMessage {
  id: string;
  type: 'file';
  quoteToken?: string;
  fileName: string;
  fileSize: number;
}

export interface LineLocationMessage {
  id: string;
  type: 'location';
  quoteToken?: string;
  title: string;
  address: string;
  latitude: number;
  longitude: number;
}

export interface LineUser {
  type: 'user';
  userId: string;
}

export interface LineGroup {
  type: 'group';
  groupId: string;
  userId?: string;
}

export interface LineRoom {
  type: 'room';
  roomId: string;
  userId?: string;
}